requirements.txt
advertools==0.13.2
amqp==5.1.1
asgiref==3.6.0
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
backcall==0.2.0
billiard==3.6.4.0
black==23.3.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
comm==0.1.3
constantly==15.1.0
contourpy==1.0.7
crispy-bootstrap5==0.7
cron-descriptor==1.4.0
cryptography==40.0.2
cssselect==1.2.0
cycler==0.11.0
dacite==1.8.1
daphne==4.0.0
debugpy==1.6.7
decorator==5.1.1
dill==0.3.6
Django==4.2
django-appconf==1.0.5
django-celery-beat==2.5.0
django-cors-headers==3.14.0
django-crispy-forms==2.0
django-filter==23.1
django-redis==5.2.0
django-select2==8.1.1
django-timezone-field==5.0
executing==1.2.0
filelock==3.12.0
flower==2.0.0
fonttools==4.39.4
htmlmin==0.1.12
humanize==4.6.0
hyperlink==21.0.0
idna==3.4
ImageHash==4.3.1
importlib-metadata==6.6.0
importlib-resources==5.12.0
incremental==22.10.0
ipykernel==6.23.0
ipython==8.13.2
itemadapter==0.8.0
itemloaders==1.1.0
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jupyter_client==8.2.0
jupyter_core==5.3.0
kiwisolver==1.4.4
kombu==5.2.4
lxml==4.9.2
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
msgpack==1.0.5
multimethod==1.9.1
mypy-extensions==1.0.0
nest-asyncio==1.5.6
networkx==3.1
numpy==1.23.5
oauthlib==3.2.2
packaging==23.1
pandas==1.5.3
parsel==1.8.1
parso==0.8.3
pathspec==0.11.1
patsy==0.5.3
phik==0.12.3
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.0
plotly==5.14.1
prometheus-client==0.17.0
prompt-toolkit==3.0.38
Protego==0.2.1
psutil==5.9.5
pure-eval==0.2.2
pyarrow==11.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.8
PyDispatcher==2.0.7
Pygments==2.15.1
pyOpenSSL==23.1.1
pyparsing==3.0.9
python-crontab==2.7.1
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
PyWavelets==1.4.1
pywin32==306
PyYAML==6.0
pyzmq==25.0.2
queuelib==1.6.2
redis==4.5.5
requests==2.29.0
requests-file==1.5.1
requests-oauthlib==1.3.1
scipy==1.10.1
Scrapy==2.8.0
seaborn==0.12.2
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
statsmodels==0.14.0
tangled-up-in-unicode==0.2.0
tenacity==8.2.2
tldextract==3.4.1
tomli==2.0.1
tornado==6.3.1
tqdm==4.65.0
traitlets==5.9.0
Twisted==22.10.0
twisted-iocpsupport==1.0.3
twython==3.9.1
txaio==23.1.1
typeguard==2.13.3
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
validators==0.20.0
vine==5.0.0
visions==0.7.5
w3lib==2.1.1
wcwidth==0.2.6
wordcloud==1.9.2
ydata-profiling==4.2.0
zipp==3.15.0
zope.interface==6.0
